Move types to events.types.ts
